+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * New `TextShow` instances:
   * `directory`: `XdgDirectory` (from `System.Directory`)
   * `time`: `UniversalTime` (from `Data.Time.Clock`)
+  * `Win32`: Lots of datatypes added in `Win32-2.5.0.0`
 * Remove most monomorphic `TextShow` functions, as their utility is questionable, and their maintenance burden is undeniable
 * Made the instances in `TextShow.Data.Bifunctor` poly-kinded when possible
 * Fix testsuite compilation on older GHCs